Move files to another package
