+++ bazel/src/main/kotlin/org/virtuslab/bazelsteward/bazel/BUILD.bazel
