+++ bazel/src/main/kotlin/org/virtuslab/bazelsteward/bazel/BazelVersion.kt
@@ -1,4 +1,4 @@
-package bazel.src
+package org.virtuslab.bazelsteward.bazel
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
