+++ bazel/src/test/kotlin/org/virtuslab/bazelsteward/bazel/BUILD.bazel
@@ -2,6 +2,6 @@ unit_tests(
     srcs = glob(["**/*.kt"]),
     test_package = "org.virtuslab.bazelsteward.bazel",
     deps = [
-        "//bazel/src",
+        "//bazel",
     ],
 )
